requirements.txt
Flask==2.3.3
twilio==8.5.0
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
asyncssh==2.14.2
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2024.6.2
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.2.35
cryptography==40.0.1
curlify==2.2.1
db-dtypes==1.1.1
decorator==5.1.1
Deprecated==1.2.13
Django==4.2
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.33.3
executing==1.2.0
facebook_business==21.0.1
flower==1.2.0
forex-python==1.8
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gcsfs==2023.3.0
geocoder==1.38.1
google-ads==25.1.0
google-api-core==2.21.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-bigquery==3.9.0
google-cloud-bigquery-storage==2.19.1
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-currency==1.0.10
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.67.0
grpcio-status==1.62.3
gunicorn==23.0.0
httplib2==0.22.0
humanize==4.6.0
idna==3.7
ipython==8.12.0
jedi==0.18.2
json-flatten==0.3
kombu==5.2.4
lxml==4.9.2
mailchimp-transactional==1.0.56
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.24.0
protobuf==4.25.5
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyactiveresource==2.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycountry==22.3.5
pycparser==2.21
Pygments==2.18.0
PyJWT==2.6.0
pymongo==4.7.3
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
ratelim==0.1.6
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ShopifyAPI==12.4.0
simplejson==3.18.4
six==1.16.0
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.6.2
text-unidecode==1.3
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.12.2
tzdata==2023.3
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
yarl==1.8.2
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
openai==1.3.8
python-multipart==0.0.9
jinja2==3.1.4
annotated-types==0.7.0
anyio==4.4.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.4.0
orjson==3.10.4
pydantic_core==2.18.4
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
ujson==5.10.0
watchfiles==0.22.0
websockets==12.0
